Move Application class to project root and rename it accordingly
